🎉 Initial commit
This commit initializes the repository with my Hugo theme, Toph. Most of
the commit history for this work could be found in my personal website.
However, due to my need for making a new website, it was finally time to
create a standalone repository for the theme. Going forward, for my own
projects with Hugo and this theme, I will clone it as a submodule and
focus on making contributions in this standalone repository.

The only changes made from the GitLab repository were in the README file
to better document the new installation instructions. They were actually
already written in anticipation of one day making this change, so I just
uncommented the instructions that were already there.

Signed-off-by: Justin Wheeler <git@jwheel.org>